backend/pyproject.toml
[tool.poetry]
name = "hospital-access-control"
version = "0.1.0"
description = "Final Year Project"
authors = ["ibrooz94 <ibrooz94@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"
joserfc = "^0.9.0"
python-multipart = "^0.0.9"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
emails = "^0.6"
jinja2 = "^3.1.3"
alembic-postgresql-enum = "^1.1.2"
pytest = "^8.1.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
fastapi-filter = {extras = ["sqlalchemy"], version = "^2.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
